requirements.txt
Pillow==10.0.1
piexif==1.1.3
click==8.1.7

